recoil sampleをコンポーネント化
